requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 科学计算
numpy>=1.24.0
scipy>=1.10.0

# 图像处理
Pillow>=9.5.0
opencv-python>=4.7.0
scikit-image>=0.20.0

# 小波变换
PyWavelets>=1.4.0

# 数据处理和可视化
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.13.0

# 其他工具
pyyaml>=6.0
tensorboardX>=2.6
